requirements.txt
astroid==2.4.0
Django==2.2
django-filter==2.2.0
djangorestframework==3.11.0
isort==4.3.21
lazy-object-proxy==1.4.3
Markdown==3.2.1
mccabe==0.6.1
pylint==2.5.0
pytz==2020.1
six==1.14.0
sqlparse==0.3.1
toml==0.10.0
typed-ast==1.4.1
wrapt==1.12.1
